// src/pages/PaymentRecord/typings.d.ts
// @ts-ignore
/* eslint-disable */

declare namespace API {
  type Params = {
    companyName: string,
    operateType: number,
    orderNo: string,
    pageNumber: number,
    pageSize: number
  };

  type TableItem = {
    invoiceType: string,
    operateType: string,
    orderAmount: number,
    orderId: number,
    orderNo: string,
    payImgUrl: string,
    payeeCompanyAccountNo: string,
    payeeCompanyId: number,
    payeeCompanyName: string,
    paymentAmount: number,
    paymentDate: string,
    paymentId: number,
    paymentType: string,
    payorCompanyAccountNo: string,
    payorCompanyId: number,
    payorCompanyName: string,
    remark: string,
    taxAmount: number,
    taxRate: number
  }
}